examples/a2a_demo/requirements.txt
a2a-sdk[http-server]
httpx
uvicorn
